pyproject.toml
[project]
name = "switchly"
version = "0.9.0"
description = "Switchly gives you runtime control of your APIs to toggle features, schedule maintenance, enforce rate limits, and perform rollouts without redeploying."
readme = "README.md"
license = { text = "MIT" }
requires-python = ">=3.11"
authors = [
    { name = "Richard Quaicoe" },
]
keywords = ["fastapi", "api", "maintenance", "middleware", "route", "lifecycle"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Framework :: FastAPI",
    "Topic :: Internet :: WWW/HTTP :: HTTP Servers",
    "Topic :: Software Development :: Libraries :: Python Modules",
]
dependencies = [
    "pydantic>=2.0",
    "anyio>=4.0",
    "starlette>=0.40",
]

[project.urls]
Homepage = "https://github.com/Attakay78/switchly"
Repository = "https://github.com/Attakay78/switchly"
Issues = "https://github.com/Attakay78/switchly/issues"

[project.optional-dependencies]
fastapi   = ["fastapi>=0.115"]
redis     = ["redis[asyncio]>=5.0"]
dashboard = [
    "jinja2>=3.1",
 "aiofiles>=23.0",
 "python-multipart>=0.0.22",
]
cli       = ["typer>=0.12", "httpx>=0.27", "rich>=13.0"]
admin     = [
    "jinja2>=3.1",
    "aiofiles>=23.0",
    "python-multipart>=0.0.22",
]
yaml      = ["pyyaml>=6.0"]
toml      = ["tomli-w>=1.0"]
rate-limit = ["limits>=5.8.0"]
flags     = [
    "openfeature-sdk>=0.8",
    "packaging>=23.0",
]
all       = [
    "fastapi>=0.115",
    "redis[asyncio]>=5.0",
    "jinja2>=3.1",
    "aiofiles>=23.0",
    "typer>=0.12",
    "httpx>=0.27",
    "rich>=13.0",
    "pyyaml>=6.0",
    "tomli-w>=1.0",
    "python-multipart>=0.0.22",
    "limits>=5.8.0",
    "openfeature-sdk>=0.8",
    "packaging>=23.0",
]
docs = [
    "mkdocs-material>=9.5",
    "mkdocstrings[python]>=0.25",
    "mkdocs-git-revision-date-localized-plugin>=1.2",
]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.23",
    "httpx>=0.27",
    "anyio[trio]",
    "ruff",
    "mypy",
    "aiofiles>=23.0",
    "fastapi>=0.115",
    "pre-commit>=3.7",
    "tomli-w>=1.0",
    "pyyaml>=6.0",
    "limits>=5.8.0",
    "openfeature-sdk>=0.8",
    "packaging>=23.0",
    "mkdocs-material>=9.5",
    "mkdocstrings[python]>=0.25",
    "mkdocs-git-revision-date-localized-plugin>=1.2",
]

[project.scripts]
switchly = "switchly.cli.main:cli"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["switchly"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I", "UP"]

[tool.ruff.lint.isort]
known-first-party = ["switchly"]

[tool.ruff.lint.per-file-ignores]
"switchly/core/feature_flags/__init__.py" = ["E402"]
"switchly/core/feature_flags/hooks.py" = ["E402"]
"switchly/core/feature_flags/provider.py" = ["E402"]

[tool.mypy]
python_version = "3.11"
strict = true

[dependency-groups]
dev = [
    "mkdocs-git-revision-date-localized-plugin>=1.2",
    "mkdocs-material>=9.5",
    "mkdocstrings[python]>=0.25",
]

